Renamed receiver package to handler
